requirements.txt
requests>=2.32.3
pydub>=0.25.1
numpy>=2.1.0
pysilero-vad==2.0.0
faster-whisper==1.0.3
mock.gpio~=0.1.8
silero-tts==0.0.4
llama-cpp-python==0.2.90